requirements.txt
tensorflow-macos==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.37.1
numpy==1.24.3
opencv-python
Pillow
ffmpeg-python
audioread==3.0.1
librosa==0.11.0
scipy==1.15.3
soundfile==0.13.1
